First Lab Question Book
The PDF/text is a collection of six programming problems related to pointers in C++. The problems cover topics such as storing information in arrays, using pointers to find sum, difference, product, square, and mean of integers, computing the area of a rectangle, and finding the largest and smallest number from user input. This is part of a lab for an Object Oriented Programming course.